+++ web/src/components/common.js
@@ -1,0 +1,9 @@
+export default {
+    downloadFile: function(contents, filename, mimetype) { // eslint-disable-line
+        let link = document.createElement("a");
+        let blob = new Blob([contents], {type: mimetype});
+        link.href = URL.createObjectURL(blob);
+        link.download = filename;
+        link.click();
+    }
+}
